+++ src/main/java/chess/ChessApplication.java
@@ -18,9 +18,12 @@ import chess.dao.RoomDao;
 import chess.dao.SquareDao;
 import chess.dto.ResponseDto;
 import chess.dto.ScoreDto;
-import chess.model.Board;
+import chess.model.board.Board;
+import chess.model.member.Member;
+import chess.model.room.Room;
 import chess.model.piece.Piece;
 import chess.model.square.Square;
+import chess.service.ChessService;
 import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
