+++ src/main/java/chess/model/ChessCommand.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.model;
 
 import java.util.Arrays;
 import java.util.List;
