+++ src/main/java/chess/model/member/Member.java
@@ -1,4 +1,4 @@
-package chess;
+package chess.model.member;
 
 public class Member {
