+++ src/main/java/chess/model/piece/PointMovingPiece.java
@@ -1,6 +1,6 @@
 package chess.model.piece;
 
-import chess.model.ConsoleBoard;
+import chess.model.board.ConsoleBoard;
 import chess.model.square.Square;
 import java.util.Collections;
 import java.util.List;
