+++ src/test/java/chess/model/piece/PawnTest.java
@@ -10,7 +10,7 @@ import static chess.model.piece.Fixtures.E5;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.model.ConsoleBoard;
+import chess.model.board.ConsoleBoard;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
