+++ src/test/java/chess/model/piece/RookTest.java
@@ -12,7 +12,7 @@ import static chess.model.piece.Fixtures.H4;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import chess.model.ConsoleBoard;
+import chess.model.board.ConsoleBoard;
 import chess.model.square.File;
 import chess.model.square.Rank;
 import chess.model.square.Square;
